web-animations/animation-group/animation-group-constructor.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationGroup constructor test</title>
<meta name="assert" content="Creates a new AnimationGroup object">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-ctor-AnimationGroup-sequence-AnimationNode--children--unrestricted-double-or-AnimationTimingInput--timing">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>

var DEFAULT_TIMING = {
    delay: 0,
    endDelay: 0,
    fill: 'auto',
    iterationStart: 0.0,
    iterations: 1.0,
    duration: 'auto',
    playbackRate: 1.0,
    direction: 'normal',
    easing: 'linear'
};

function assert_timing_equlas(actual, expected){
    assert_equals(actual.delay, expected.delay, 'Value of AnimationNode.timing.delay ' +
        'is unexpected');
    assert_equals(actual.endDelay, expected.endDelay, 'Value of AnimationNode.timing.endDelay ' +
        'is unexpected');
    assert_equals(actual.fill, expected.fill, 'Value of AnimationNode.timing.fill ' +
        'is unexpected');
    assert_equals(actual.iterationStart, expected.iterationStart, 'Value of ' +
        'AnimationNode.timing.iterationStart is unexpected');
    assert_equals(actual.iterations, expected.iterations, 'Value of ' +
        'AnimationNode.timing.iterations is unexpected');
    assert_equals(actual.duration, expected.duration, 'Value of AnimationNode.timing.duration ' +
        'is unexpected');
    assert_equals(actual.playbackRate, expected.playbackRate, 'Value of ' +
        'AnimationNode.timing.playbackRate is unexpected');
    assert_equals(actual.direction, expected.direction, 'Value of ' +
        'AnimationNode.timing.direction is unexpected');
    assert_equals(actual.easing, expected.easing, 'Value of AnimationNode.timing.easing ' +
        'is unexpected');
}

test(function() {
    var group = new AnimationGroup(null);

    assert_array_equals(group.children, [], 'AnimationGroup should have no children');
    assert_timing_equals(group.timing, DEFAULT_TIMING);
}, 'Test constructor AnimationGroup(null) creates AnimationGroup instance ' +
    'without children and with default timings');

test(function() {
    var group = new AnimationGroup([]);

    assert_array_equals(group.children, [], 'AnimationGroup should have no children');
    assert_timing_equals(group.timing, DEFAULT_TIMING);
}, 'Test AnimationGroup constructor with emplty child list and with no second argument ' +
    'creates AnimationGroup instance without children and with default timings');

test(function() {
    var group = new AnimationGroup([], 5);

    assert_array_equals(group.children, [], 'AnimationGroup should have no children');
    var timing = Object.create(DEFAULT_TIMING);
    timing.duration = 5;
    assert_timing_equals(group.timing, timing);
}, 'Test AnimationGroup constructor with emplty child list and with double second argument ' +
    'creates AnimationGroup instance without children and with default timings, except duration');

test(function() {
    var timing = {
        delay: 1,
        endDelay: 1,
        fill: 'none',
        iterationStart: 1.1,
        iterations: 2.0,
        duration: 3,
        playbackRate: -2,
        playbackDirection: 'reverse',
        direction: 'reverse',
        easing: 'ease'};
    var node = new AnimationGroup([], timing);

    assert_array_equals(node.children, [], 'AnimationGroup should have no children');
    assert_timing_equals(node.timing, timing);
}, 'Test AnimationGroup constructor with emplty child list and with AnimationTimingInput ' +
    'second argument creates AnimationGroup instance without children and with given timings');

test(function() {
    var node = newAnimation(createDiv(this));
    var children = [node];
    var group = new AnimationGroup(children);

    assert_array_equals(group.children, children, 'AnimationGroup should have single child');
}, 'Test AnimationGroup constructor creates AnimationGroup instance with single child');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationGroup([]);
    var node3 = newAnimation(createDiv(this));
    var children = [node1, node2, node3];
    var group = new AnimationGroup(children);

    assert_array_equals(group.children, children, 'AnimationGroup.children should correspond ' +
        'the first constructor argument');
}, 'Test AnimationGroup constructor creates AnimationGroup instance with children');

test(function() {
    var node = newAnimation(createDiv(this));

    assert_throws('HierarchyRequestError', function() {
        var group = new AnimationGroup([node, node]);
    }, 'HierarchyRequestError should be thrown in case of attempt to create ' +
        'AnimationGroup with the same nodes');
}, 'HierarchyRequestError is thrown if child list contains the same nodes');

test(function() {
    var node = new AnimationGroup([]);

    assert_throws('HierarchyRequestError', function() {
        var group = new AnimationGroup([node, node]);
    }, 'HierarchyRequestError should be thrown in case of attempt to create ' +
        'AnimationGroup with the same nodes');
}, 'HierarchyRequestError is thrown if child list contains the same AnimationGroup nodes');
</script>
</body>
